Buildfixes for openwrt

+++ generator.cpp
@@ -97,7 +97,7 @@ Generator::send(char *data, int datalen) {
 	int res;
 	if((res = sendto(sockraw, generator_packet, datalen + sizeof(struct iphdr2), 0, (struct sockaddr *)&dest_addr, sizeof(struct sockaddr))) == -1)
 	{
-		printf("msglen[%lu]\n", datalen + sizeof(struct iphdr2));
+		printf("msglen[%lu]\n", (unsigned long) (datalen + sizeof(struct iphdr2)));
 		perror("sendto");
 	}
 	return res;
